Rename ConvertToBlahBlahBlah with just one object to simply 'Convert'.
This is for Karl's probuilder benefit -- he's using reflection, and was picking
up the wrong override.

Also mucked a bit with the control flow of the version with arrays, which had
subtly different handling depending on which menu we were coming from. Now we
get the same call to the underlying API either way.