add function to validate bones
a transform is a bone if it is in the skinned meshes's bone list or has
both a parent and child that is in the skinned meshes' list of bones.
Go through all bones and their hierarchies, creating fbx node's for them
as we go and storing the valid bones.