uni-42394: don't reset transforms when transfering anim
Viktoria pointed out this piece of code; we were resetting the transform to the identity for the path from the root being exported down to the root being animated.

That's made to be bad by the prior commit (and it never quite did what we intended), so I removed that code.

It's still possible to reset transforms when exporting but only via the API right now. Maybe we shouldn't?