uni-40523: defining all the cases and making them work
The spec wasn't clear about what would happen in which case. Now it should be, and unit tests make sure it's all happening.

Particular changes:
- now Convert returns the prefab, not the instance. That's because sometimes there isn't an instance. I had to change a few unit tests.
- if what we're converting is an fbx or an instance of an fbx, use it. That was mostly implemented before.
- if what we're converting is a prefab, update the prefab
- added unit tests to each part so we can make sure the complicated logic is properly implemented
- probably fixed a bug in file paths for prefabs in some cases; I didn't test whether I could actually trigger it

I also changed the ExporterTestBase ExportToFbx function to return the asset. Almost every call to that function wanted the asset rather than what it used to return which is its filename.