UT-109: package that actually loads (when you manually delete the editor tests)
Replaces the compiled dll with source. To have this run in the assets folder,
you need to add the COM_UNITY_FORMATS_FBX_AS_ASSET define.

The editor tests cause a crash on load. Delete them, and this package now works
as a package (embedded or sent to the cloud).

Includes the new, smaller UnityFbxSdkNative bundle on OSX for faster upload times.