UT-511: fix a bug and some tests
The auto-updater is now robust to FbxPrefab not being found, or having
duplicates, for whatever reason. In particular on 2018.1 we find it in a fixed
location because the asset database can't find anything in the packages.

Added a ChangeLog as is required; also now we use the ChangeLog to determine the version.