refactor animation only export to separate bone and other object export
the ExportGameObjectAndParents function was handling both export of bones and regular objects, which was confusing.
Separated it so the bones get exported first, then ExportGameObjectAndParents exports the remaining objects and makes sure all the objects are connected to one another.

Also added a helper for creating a single FbxNode as the code was being duplicated a few times.